feat: Implement shelf change
After a successful resolution of an update() call towards the server, getAll() is called to synch the app with the server database. This lets the server be the single source of truth for data. Otherwise, if the change is performed simultaneously on the server and in the app without synch, a page refresh might result in a change in the order of the books.