fix: Optimize shelf change
Using the server as the single source of truth resulted in sluggish interface, because rerendering occurred after the server confirmed the successful update and data was refetched.
Designating the app as the source of truth and handling the server as a storage required a method to have consistent order among the books. To achieve this, a library-type title sorting is introduced (alphabetic without trailing articles).