fix: Add check for empty query to incoming result
Because empty search query did not trigger a server call, deleting the search field one character a time resulted in the last single character being sent as query, and the correct state of empty array overwritten once the server response arrived. With a check on the prevState.searchValue the state can be correctly set to an empty array instead of using the response array.